feat: find refVersionId from findVersionQuery
